consolidate constants from controller and v2v-helper
