+++ k8s/migration/internal/controller/bmconfig_controller.go
@@ -23,7 +23,7 @@ import (
 
 	"github.com/pkg/errors"
 	vjailbreakv1alpha1 "github.com/platform9/vjailbreak/k8s/migration/api/v1alpha1"
-	constants "github.com/platform9/vjailbreak/k8s/migration/pkg/constants"
+	constants "github.com/platform9/vjailbreak/pkg/common/constants"
 	scope "github.com/platform9/vjailbreak/k8s/migration/pkg/scope"
 	providers "github.com/platform9/vjailbreak/pkg/vpwned/sdk/providers"
 	corev1 "k8s.io/api/core/v1"
