+++ k8s/migration/internal/controller/rdmdisk_controller.go
@@ -31,7 +31,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/log"
 
 	vjailbreakv1alpha1 "github.com/platform9/vjailbreak/k8s/migration/api/v1alpha1"
-	constants "github.com/platform9/vjailbreak/k8s/migration/pkg/constants"
+	constants "github.com/platform9/vjailbreak/pkg/common/constants"
 	utils "github.com/platform9/vjailbreak/k8s/migration/pkg/utils"
 	v2vutils "github.com/platform9/vjailbreak/v2v-helper/pkg/utils"
 )
