+++ k8s/migration/pkg/utils/migrationplanutils.go
@@ -18,7 +18,7 @@ import (
 	vjailbreakv1alpha1 "github.com/platform9/vjailbreak/k8s/migration/api/v1alpha1"
 	corev1 "k8s.io/api/core/v1"
 
-	"github.com/platform9/vjailbreak/k8s/migration/pkg/constants"
+	"github.com/platform9/vjailbreak/pkg/common/constants"
 	"k8s.io/apimachinery/pkg/util/validation"
 )
