+++ k8s/migration/pkg/utils/rollingmigrationutils.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/pkg/errors"
 	vjailbreakv1alpha1 "github.com/platform9/vjailbreak/k8s/migration/api/v1alpha1"
-	"github.com/platform9/vjailbreak/k8s/migration/pkg/constants"
+	"github.com/platform9/vjailbreak/pkg/common/constants"
 	scope "github.com/platform9/vjailbreak/k8s/migration/pkg/scope"
 	providers "github.com/platform9/vjailbreak/pkg/vpwned/sdk/providers"
 	"github.com/vmware/govmomi/find"
